+++ src/common/geometry.py
