+++ src/gp_server/tests/test_api_clean_paths.py
@@ -2,7 +2,7 @@ from typing import Callable, Tuple
 import json
 import pytest
 from shapely.geometry import LineString
-from gp_server.utils.geometry import project_geom
+from common.geometry import project_geom
 from gp_server.app.constants import cost_prefix_dict, TravelMode, RoutingMode
